Changed all salary and bonus related fields from int to double

+++ InHumanResources/IHR/src/main/java/com/ihr/ihr/common/dtos/BonusDtos/CreateBonusDto.java
@@ -4,7 +4,7 @@ public class CreateBonusDto extends UpdateBonusDto {
 
     Long idPayment;
 
-    public CreateBonusDto(Integer value, String bonusDescription, Long idPayment) {
+    public CreateBonusDto(Double value, String bonusDescription, Long idPayment) {
         super(value, bonusDescription);
         this.idPayment = idPayment;
     }
