+++ InHumanResources/IHR/src/main/java/com/ihr/ihr/common/interf/BonusValidation.java
@@ -1,10 +1,9 @@
 package com.ihr.ihr.common.interf;
 
-import com.ihr.ihr.common.dtos.BonusDtos.BonusDto;
 import com.ihr.ihr.common.dtos.BonusDtos.UpdateBonusDto;
 
 public interface BonusValidation {
-    boolean isValueValid(Integer value);
+    boolean isValueValid(Double value);
     boolean isDenumireBonusValid(String denumireBonus);
     boolean isBonusDtoValid(UpdateBonusDto bonusDto);
 }
